changes from today's lab

+++ list.h
@@ -9,13 +9,13 @@ typedef struct header {
 //    boolean free;
     //look up type for single bit for free or not... //TODO
     //TODO: pad to 32 bytes here and then set check sums
-    int firstCheckSum;
+    short firstCheckSum;
     boolean free;
     struct header *nextHeader; //first linked list
     struct header *nextFree; //second linked list
     long amountAllocated; //to be used for compaction
     // know what the user has because you round up to nearest multiple of 8
-    int secondCheckSum;
+    short secondCheckSum;
     //TODO: add check sums
 } header;
